Adding a prototype Jupyter notebook
This notebook documents an atmospheric dycore verification test: The
Shallow Water Test Case 1. Its purpose is to demonstrate how a Jupyter
notebook can be used to describe a verification test and to report on
the results of nightly (or regularly scheduled) testing.

This notebook is intended to provide a concrete example of the
literate programming techniques proposed in the Verification task of
the CMDV Software proposal.  Moreover, it is intended to spur
discussion among the CMDV Software Verification team about what
components belong in the verification test documents, and ultimately
lead to a template for use in Jupyter notebooks for the atmospheric
physics parameterizations promised in the proposal.