Updated prototype to support alternative citation syntax
The first version of converter.py worked for markdown citations of the
form [@bibtex_key]. Now it supports ciations of the following forms as
well:

    [@bibtex_key1; @bibtex_key2]
    [see @bibtex_key]
    [@bibtex_key, pp. 12-34]
    [see @bibtex_key, pp. 123-456]
    @bibtex_key
    [-@bibtex_key]

These forms were already supported by pandoc and pandoc-citeproc,
which are used within converter.py, so this update makes this system
compatible with all uses of pandoc.

Other changes this commit: SimpleExample changed to CitationTest, and
ref.bib expanded to support more references.