Updated converter.py to be configurable
Configuration improvements start with the AddCitationsPreprocessor
class, which can now successfully be configured with a Config object,
making it compatible with all other Jupyter and nbconvert objects.
The converter.py script itself was also updated to accept command-line
arguments that can be passed to the AddCitationsPreprocessor
object. The following attributes are now configurable:

    o Verbosity
    o Citation Style Language file
    o BibTeX bibliography database file
    o Header title for the bibliography section

Also, converter.py can now work on multiple Jupyter notebooks
specified on a single command-line.