Fixes to cmake/cdash scripts.
Copying lib directory to build tree, so that python scripts therein
can be modified based on cmake variables.

Adding script which modifies CMDV.py to set the full path to clone-dir.sh, which
is required with the cmake/cdash workflow.

Still need to modify path in one more script to get tests to run to completion.