Removed version specifications from environment files
The purpose of doing this is to increase compatibility on Mac and
Linux (and potentially other) platforms.

The py3 environment works for me (after applying the `webbrowser.py`
workaround: see https://github.com/jupyter/notebook/issues/3746).

The py2 environment results in a server error for me, although I
suspect it is not related to my changes.  I'd be curious to see how it
works on Linux.